backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
asyncpg
pydantic-settings
python-jose
passlib[bcrypt]
